Adding my changes to the navbar
Right now it's not working. The sidebar does correctly announce an event and, when it was displaying, the top bar did receive the event. I was trying to restructure the way the top bar was defined and at some point it just stopped showing at all even through the middleware runs fine.